[2747] Integration test 16 update
