+++ Compiler/TranslatorTests/TestProjects/16/Bridge/reference/testProject.meta.js
@@ -3,7 +3,7 @@ Bridge.assembly("TestProject", function ($asm, globals) {
 
 
     var $m = Bridge.setMetadata,
-        $n = [System,Test.BridgeIssues.N059,Test.BridgeIssues.N065,Test.BridgeIssues.N064,Test.BridgeIssues.N475A,Test.BridgeIssues.N883A,Test.BridgeIssues.N882,System.Collections.Generic,Test.BridgeIssues.N1023,Test.BridgeIssues.N1552,Test.BridgeIssues.N2420,Test.BridgeIssues.N2318,Test.BridgeIssues.N2143,Test.BridgeIssues.N384,Test.BridgeIssues.N089,Test.BridgeIssues.N084,Test.BridgeIssues.N066,Test.BridgeIssues.N062,Test.BridgeIssues.N391,Test.BridgeIssues.N447,Test.BridgeIssues.N694,Test.BridgeIssues.N557,Test.BridgeIssues.N542,Test.BridgeIssues.N534,Test.BridgeIssues.N524,Test.BridgeIssues.N477,Test.BridgeIssues.N411,Test.BridgeIssues.N475B,Test.BridgeIssues.N963,Test.BridgeIssues.N883,Test.BridgeIssues.N856,Test.BridgeIssues.N1499,Test.BridgeIssues.N1472,Test.BridgeIssues.N1461,Test.BridgeIssues.N1412,Test.BridgeIssues.N1215,Test.BridgeIssues.N1212,Test.BridgeIssues.N1151,Test.BridgeIssues.N1009,Test.BridgeIssues.N1537,Test.BridgeIssues.N1508,Test.BridgeIssues.N2277,Test.BridgeIssues.N2113,Test.BridgeIssues.N2567];
+        $n = [System,Test.BridgeIssues.N059,Test.BridgeIssues.N065,Test.BridgeIssues.N064,Test.BridgeIssues.N475A,Test.BridgeIssues.N883A,Test.BridgeIssues.N882,System.Collections.Generic,Test.BridgeIssues.N1023,Test.BridgeIssues.N1552,Test.BridgeIssues.N2420,Test.BridgeIssues.N2318,Test.BridgeIssues.N2143,Test.BridgeIssues.N384,Test.BridgeIssues.N089,Test.BridgeIssues.N084,Test.BridgeIssues.N066,Test.BridgeIssues.N062,Test.BridgeIssues.N391,Test.BridgeIssues.N447,Test.BridgeIssues.N694,Test.BridgeIssues.N557,Test.BridgeIssues.N542,Test.BridgeIssues.N534,Test.BridgeIssues.N524,Test.BridgeIssues.N477,Test.BridgeIssues.N411,Test.BridgeIssues.N475B,Test.BridgeIssues.N963,Test.BridgeIssues.N883,Test.BridgeIssues.N856,Test.BridgeIssues.N1499,Test.BridgeIssues.N1472,Test.BridgeIssues.N1461,Test.BridgeIssues.N1412,Test.BridgeIssues.N1215,Test.BridgeIssues.N1212,Test.BridgeIssues.N1151,Test.BridgeIssues.N1009,Test.BridgeIssues.N1537,Test.BridgeIssues.N1508,Test.BridgeIssues.N2277,Test.BridgeIssues.N2113,Test.BridgeIssues.N2567,Test.BridgeIssues.N2747];
     $m($n[1].Class59, function () { return {"att":1048576,"a":4,"m":[{"a":2,"isSynthetic":true,"n":".ctor","t":1,"sn":"ctor"},{"a":2,"n":"Method1","is":true,"t":8,"sn":"Method1","rt":$n[0].Void},{"a":2,"n":"Method1","t":8,"pi":[{"n":"d","pt":$n[1].Class59.Aux1,"ps":0}],"sn":"Method1","rt":$n[0].Void,"p":[$n[1].Class59.Aux1]}]}; });
     $m($n[1].Class59.Aux1, function () { return {"td":$n[1].Class59,"att":1048578,"a":2,"m":[{"a":2,"isSynthetic":true,"n":".ctor","t":1,"sn":"ctor"}]}; });
     $m($n[14].Class89, function () { return {"att":1048577,"a":2,"m":[{"a":2,"isSynthetic":true,"n":".ctor","t":1,"sn":"ctor"},{"a":1,"n":"Test","t":8,"pi":[{"n":"p","ip":true,"pt":$n[0].Array.type(System.Object),"ps":0}],"sn":"Test","rt":$n[0].Void,"p":[$n[0].Array.type(System.Object)]}]}; });
@@ -81,6 +81,7 @@ Bridge.assembly("TestProject", function ($asm, globals) {
     $m($n[12].Level1$1.Level2.Level3, function () { return {"td":$n[12].Level1$1.Level2,"att":1048578,"a":2,"m":[{"a":2,"isSynthetic":true,"n":".ctor","t":1,"sn":"ctor"},{"a":2,"n":"Value","t":16,"rt":System.Object,"g":{"a":2,"n":"get_Value","t":8,"rt":System.Object,"fg":"Value"},"fn":"Value"}]}; });
     $m($n[42].Bridge2113, function () { return {"att":1048577,"a":2,"m":[{"a":2,"isSynthetic":true,"n":".ctor","t":1,"sn":"ctor"},{"a":2,"n":"Test","is":true,"t":8,"sn":"Test","rt":$n[0].String}]}; });
     $m($n[43].N2567, function () { return {"att":1048577,"a":2,"m":[{"a":2,"isSynthetic":true,"n":".ctor","t":1,"sn":"ctor"},{"a":2,"n":"CheckAllComments","is":true,"t":8,"sn":"CheckAllComments","rt":$n[0].Void},{"a":2,"n":"CheckCommentsInInitPositionAfter","is":true,"t":8,"sn":"CheckCommentsInInitPositionAfter","rt":$n[0].Void},{"a":2,"n":"CheckCommentsInInitPositionBefore","is":true,"t":8,"sn":"CheckCommentsInInitPositionBefore","rt":$n[0].Void},{"a":2,"n":"CheckCommentsInInitPositionBottom","is":true,"t":8,"sn":"CheckCommentsInInitPositionBottom","rt":$n[0].Void},{"a":2,"n":"CheckCommentsInInitPositionTop","is":true,"t":8,"sn":"CheckCommentsInInitPositionTop","rt":$n[0].Void},{"a":2,"n":"CheckMultilineInjectComment","is":true,"t":8,"sn":"CheckMultilineInjectComment","rt":$n[0].Void},{"a":2,"n":"CheckMultilineInjectCommentAsterisk","is":true,"t":8,"sn":"CheckMultilineInjectCommentAsterisk","rt":$n[0].Void},{"a":2,"n":"CheckMultilineInjectCommentWithCode","is":true,"t":8,"sn":"CheckMultilineInjectCommentWithCode","rt":$n[0].Void},{"a":2,"n":"CheckSinglelineInjectComment","is":true,"t":8,"sn":"CheckSinglelineInjectComment","rt":$n[0].Void},{"a":2,"n":"CheckTheSameLineAndEmptyComments","is":true,"t":8,"sn":"CheckTheSameLineAndEmptyComments","rt":$n[0].Void},{"a":2,"n":"CheckTheSameLineAndEmptyInjectComments","is":true,"t":8,"sn":"CheckTheSameLineAndEmptyInjectComments","rt":$n[0].Void}]}; });
+    $m($n[44].N2747, function () { return {"att":1048577,"a":2,"m":[{"a":2,"isSynthetic":true,"n":".ctor","t":1,"sn":"ctor"},{"a":2,"n":"Output","is":true,"t":8,"sn":"Output","rt":$n[0].Void}]}; });
     $m($n[13].Person, function () { return {"att":1048577,"a":2,"m":[{"a":2,"isSynthetic":true,"n":".ctor","t":1,"sn":"ctor"}]}; });
     $m($n[13].N384, function () { return {"att":1048576,"a":4,"m":[{"a":2,"isSynthetic":true,"n":".ctor","t":1,"sn":"ctor"}]}; });
     $m($n[13].N384.App, function () { return {"td":$n[13].N384,"att":1048578,"a":2,"m":[{"a":2,"isSynthetic":true,"n":".ctor","t":1,"sn":"ctor"},{"a":2,"n":"Main1","is":true,"t":8,"sn":"Main1","rt":$n[0].Void}]}; });
