+++ Compiler/TranslatorTests/TestProjects/16/test.csproj
@@ -84,8 +84,8 @@
     <Compile Include="BridgeIssues\2500\N2277.cs" />
     <Compile Include="BridgeIssues\2500\N2143.cs" />
     <Compile Include="BridgeIssues\2500\N2113.cs" />
-	<Compile Include="BridgeIssues\2500\N2747.cs" />
     <Compile Include="BridgeIssues\3000\N2567.cs" />
+    <Compile Include="BridgeIssues\3000\N2747.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="BridgeIssues\0500\N384.cs" />
   </ItemGroup>
